Fix final compilation errors: CommentController type casting and MessageService Kafka parameters

+++ spring_boot_backend/src/main/kotlin/com/androidinsta/controller/User/CommentController.kt
@@ -37,7 +37,7 @@ class CommentController(
             parentCommentId = request.parentCommentId
         )
 
-        val repliesCount = commentService.getRepliesCount(comment.id)
+        val repliesCount = commentService.getRepliesCount(comment.id).toInt()
 
         val response = CommentResponse(
             id = comment.id,
@@ -64,7 +64,7 @@ class CommentController(
         val comments = commentService.getPostComments(postId)
 
         val responses = comments.map { comment ->
-            val repliesCount = commentService.getRepliesCount(comment.id)
+            val repliesCount = commentService.getRepliesCount(comment.id).toInt()
             
             CommentResponse(
                 id = comment.id,
